scripts/requirements.txt
python-box[all]~=6.0
jsonschema
